Make map functions as reusable helpers
